Augmented Inverse Probablity Weighting (AIPW)

Define an R6Class aipw object

Format

R6Class object.

Value

AIPW object

Details

create an AIPW object

Super class

AIPW::AIPW_base -> AIPW

Public fields

libs

SuperLearner or sl3 libraries and their fitted objects

sl.fit

A wrapper function for fitting SuperLearner or sl3

sl.predict

A wrapper using sl.fit to predict

Methods

Public methods

AIPW$new()

AIPW$fit()

AIPW$clone()

Inherited methods

AIPW::AIPW_base$plot.p_score()

AIPW::AIPW_base$summary()

Method new()

Create a new AIPW object.

Usage

AIPW$new(
  Y = NULL,
  A = NULL,
  verbose = FALSE,
  W = NULL,
  W.Q = NULL,
  W.g = NULL,
  Q.SL.library = NULL,
  g.SL.library = NULL,
  k_split = 10
)

Arguments

Y

Outcome (binary integer: 0 or 1)

A

Exposure (binary integer: 0 or 1)

verbose

Whether to print the result (logical; Default = FALSE)

W

covariates for both exposure and outcome models (vector, matrix or data.frame). If null, this function will seek for inputs from W.Q and W.g.

W.Q

Only valid when W is null, otherwise it would be replaced by W. Covariates for outcome model (vector, matrix or data.frame).

W.g

Only valid when W is null, otherwise it would be replaced by W. Covariates for exposure model (vector, matrix or data.frame)

Q.SL.library

SuperLearner libraries or sl3 learner object (Lrnr_base) for outcome model

g.SL.library

SuperLearner libraries or sl3 learner object (Lrnr_base) for exposure model

k_split

Number of splitting (integer; range: from 1 to number of observation-1): if k_split=1, no sample splitting; if k_split>1, use similar technique of cross-validation (e.g., k_split=10, use 9/10 of the data to estimate and the remaining 1/10 leftover to predict) NOTE: it's recommended to use sample splitting.

Returns

A new AIPW object

Examples

library(SuperLearner)
aipw_sl &lt;- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)

Method fit()

Fitting the data into the AIPW object with/without sample splitting to estimate the influence functions

Usage

AIPW$fit()

Returns

A fitted AIPW object

Examples

library(SuperLearner)
aipw_sl &lt;- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)
aipw_sl$fit()

Method clone()

The objects of this class are cloneable with this method.

Usage

AIPW$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `AIPW$new`
## ------------------------------------------------

library(SuperLearner)

#> Loading required package: nnls

#> Super Learner

#> Version: 2.0-26

#> Package created on 2019-10-27

aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)

## ------------------------------------------------
## Method `AIPW$fit`
## ------------------------------------------------

library(SuperLearner)
aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)
aipw_sl$fit()