Augmented Inverse Probablity Weighting (AIPW) take TMLE or tmle3 inputs

Define an R6Class aipw_tmle object whcih takes TMLE or tmle3 inputs

Format

R6Class object.

Value

AIPW object

Details

create an AIPW object

Super class

AIPW::AIPW_base -> AIPW_tmle

Methods

Public methods

AIPW_tmle$new()

AIPW_tmle$clone()

Inherited methods

AIPW::AIPW_base$calculate_result()

AIPW::AIPW_base$plot.p_score()

Method new()

Create a new AIPW_tmle object.

Usage

AIPW_tmle$new(Y = NULL, A = NULL, tmle_fit = NULL, verbose = TRUE)

Arguments

Y

Outcome (binary integer: 0 or 1)

A

Exposure (binary integer: 0 or 1)

tmle_fit

A fitted tmle object

verbose

Whether to show progression bar and print the result (logical; Default = FALSE)

Returns

A new AIPW_tmle obejct

Examples

\dontrun{
vec &lt;- function() sample(0:1,100,replace = TRUE)
df &lt;- data.frame(replicate(4,vec()))
names(df) &lt;- c("A","Y","W1","W2")

## From tmle
require(tmle)
require(SuperLearner)
tmle_fit &lt;- tmle(Y=df$Y,A=df$A,W=subset(df,select=c("W1","W2")),
                 Q.SL.library="SL.glm",
                 g.SL.library="SL.glm",
                 family="binomial")
AIPW_tmle$new(A=df$A,Y=df$Y,tmle_fit = tmle_fit,verbose = TRUE)$calculate_result()

## From tmle3
# tmle3 simple implementation
require(tmle3)
require(sl3)
node_list &lt;- list(A = "A",Y = "Y",W = c("W1","W2"))
or_spec &lt;- tmle_OR(baseline_level = "0",contrast_level = "1")
tmle_task &lt;- or_spec$make_tmle_task(df,node_list)
lrnr_glm &lt;- make_learner(Lrnr_glm)
sl &lt;- Lrnr_sl$new(learners = list(lrnr_glm))
learner_list &lt;- list(A = sl, Y = sl)
tmle3_fit &lt;- tmle3(or_spec, data=df, node_list, learner_list)

# parse tmle3_fit into AIPW_tmle class
AIPW_tmle$new(A=df$A,Y=df$Y,tmle_fit = tmle3_fit,verbose = TRUE)$calculate_result()
}

Method clone()

The objects of this class are cloneable with this method.

Usage

AIPW_tmle$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `AIPW_tmle$new`
## ------------------------------------------------

if (FALSE) {
vec <- function() sample(0:1,100,replace = TRUE)
df <- data.frame(replicate(4,vec()))
names(df) <- c("A","Y","W1","W2")

## From tmle
require(tmle)
require(SuperLearner)
tmle_fit <- tmle(Y=df$Y,A=df$A,W=subset(df,select=c("W1","W2")),
                 Q.SL.library="SL.glm",
                 g.SL.library="SL.glm",
                 family="binomial")
AIPW_tmle$new(A=df$A,Y=df$Y,tmle_fit = tmle_fit,verbose = TRUE)$calculate_result()

## From tmle3
# tmle3 simple implementation
require(tmle3)
require(sl3)
node_list <- list(A = "A",Y = "Y",W = c("W1","W2"))
or_spec <- tmle_OR(baseline_level = "0",contrast_level = "1")
tmle_task <- or_spec$make_tmle_task(df,node_list)
lrnr_glm <- make_learner(Lrnr_glm)
sl <- Lrnr_sl$new(learners = list(lrnr_glm))
learner_list <- list(A = sl, Y = sl)
tmle3_fit <- tmle3(or_spec, data=df, node_list, learner_list)

# parse tmle3_fit into AIPW_tmle class
AIPW_tmle$new(A=df$A,Y=df$Y,tmle_fit = tmle3_fit,verbose = TRUE)$calculate_result()
}